// src/app/core/models/class.model.ts
export interface SessionClassModel {
  _id: string;
  sessionNumber: string;
  sessionStartDate: string;
  sessionEndDate: string;
  sessionStartTime: string;
  sessionEndTime: string;
  instructorId: string;
  students: StudentRegisteredClass[];
}
export interface ClassModel {
  status: String;
  _id: String;
  courseId: any;
  courseName: String;
  classAccessType: String;
  classDeliveryType: String;
  instructorCost: String;
  instructorCostCurrency: String;
  isGuaranteedToRun : boolean;
  externalRoom:boolean;
  minimumEnrollment: String;
  maximumEnrollment: String;
  classStartDate: String;
  classEndDate: String;
  sessions: SessionClassModel[];
}
// ...
export interface Session {
  _id: string;
  instructorId: string;
  laboratoryId: string;
  sessionEndDate: string;
  sessionEndTime: string;
  sessionNumber: number;
  sessionStartDate: string;
  sessionStartTime: string;
}
// ...
export interface Student extends Pagination {
  __v: number;
  _id: string;
  classId: ClassModel;
  courseId: CourseModel;
  createdAt: string;
  id: string;
  registeredOn: string;
  session: any[];
  status: string;
  studentId: StudenId;
  updatedAt: string;
  classStartDate: string;
  classEndDate: string;
}

export interface StudenId {
  _id: string;
  name: string;
  email: string;
  last_name: string;
  id: string;
}

export interface StudentPaginationModel extends Pagination {
  docs: Student[];
  filterText: string;
}

export interface Session {
  sessionNumber: number;
  sessionStatus:string;
}

export interface StudentApproval {
  studentId: string;
  classId: String;
  status: string;
  approvedOn: string;
  approvedBy: string;
  session: Session[];
}